+++ insar_explorer.py
@@ -192,7 +192,10 @@ class InsarExplorer:
         # for reuse if plugin is reopened
         # Commented next statement since it causes QGIS crashe
         # when closing the docked window:
-        # self.dockwidget = None
+        self.dockwidget = None
+
+        # Released memory when widget is closed
+        self.gui_controller = None
 
         self.pluginIsActive = False
 
@@ -234,7 +237,8 @@ class InsarExplorer:
             # TODO: fix to allow choice of dock location
             self.iface.addDockWidget(Qt.BottomDockWidgetArea, self.dockwidget)
             self.dockwidget.show()
-            self.gui_controller = GuiController(self)
+            if self.gui_controller is None:
+                self.gui_controller = GuiController(self)
         else: # this will reload the widget when button clicked again
             self.iface.removeDockWidget(self.dockwidget)
             self.dockwidget = None